+++ src/main/java/ru/startup/controllers/ExtremeController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.ExtremeDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class ExtremeController {
     }
 
     @GetMapping("/api/extreme/{id}")
-    public ExtremeDTO getExtremeById(@PathVariable Long id){
-        return extremeService.getExtremeById(id);
+    public ResponseEntity<ExtremeDTO> getExtremeById(@PathVariable Long id){
+        if (!extremeService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(extremeService.getExtremeById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/extreme/{id}")
-    public void deleteExtremeById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteExtremeById(@PathVariable Long id){
+        if (!extremeService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         extremeService.deleteExtremeById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/extreme")
