+++ src/main/java/ru/startup/controllers/WindTunnelController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.WindTunnelDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class WindTunnelController {
     }
 
     @GetMapping("/api/wind-tunnel/{id}")
-    public WindTunnelDTO getWindTunnelById(@PathVariable Long id){
-        return windTunnelService.getWindTunnelById(id);
+    public ResponseEntity<WindTunnelDTO> getWindTunnelById(@PathVariable Long id){
+        if (!windTunnelService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(windTunnelService.getWindTunnelById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/wind-tunnel/{id}")
-    public void deleteWindTunnelById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteWindTunnelById(@PathVariable Long id){
+        if (!windTunnelService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         windTunnelService.deleteWindTunnelById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/wind-tunnel")
