+++ src/main/java/ru/startup/controllers/ZooController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.ZooDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class ZooController {
     }
 
     @GetMapping("/api/zoo/{id}")
-    public ZooDTO getZooById(@PathVariable Long id){
-        return zooService.getZooById(id);
+    public ResponseEntity<ZooDTO> getZooById(@PathVariable Long id){
+        if (!zooService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(zooService.getZooById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/zoo/{id}")
-    public void deleteZooById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteZooById(@PathVariable Long id){
+        if (!zooService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         zooService.deleteZooById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/zoo")
